haberozet/requirements.txt
streamlit>=1.35.0
newspaper3k>=0.2.8
nltk>=3.8.1
scikit-learn>=1.4.0
networkx>=3.2.0
numpy>=1.26.0,<2
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
lxml_html_clean>=0.1.0
Pillow>=10.0.0
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
